refactor: split tx entities in multiple files
