+++ lib/core/transaction/domain/entities/tx_vout.dart
@@ -1,0 +1,16 @@
+import 'package:bb_mobile/core/transaction/domain/entities/tx_script_sig.dart';
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'tx_vout.freezed.dart';
+part 'tx_vout.g.dart';
+
+@freezed
+class TxVout with _$TxVout {
+  const factory TxVout({
+    required BigInt value,
+    required int n,
+    required TxScriptSig scriptPubKey,
+  }) = _TxVout;
+
+  factory TxVout.fromJson(Map<String, dynamic> json) => _$TxVoutFromJson(json);
+}
